Remove Module System
# Rationale

This system was introduced in early 2014 to allow the BungeeCord
repository to include more "modules" in the future instead of stuffing
them into a separate repository.

Since its introduction, the selection of modules available (such as
`cmd_alert` and `reconnect_yaml`) have not changed. It's clear that after
four years of being able to introduce new modules that the system has
not worked out as expected.

# Implementation

The module system has been totally removed. All modules (except for
`reconnect_yaml`) are now integrated into the proxy core.