Remove BungeeCord Security Manager
# Rationale

The security manager (and related machinery) was introduced in BungeeCord
in July 2014 to encourage plugins to use BungeeCord-specific coding
practices.

Since its introduction, the restrictions have never been enforced. The
security manager was unpopular in the community when introduced, and a
significant number of people have forked BungeeCord to disable this
feature, including Waterfall. Today, it only serves to nag people. It
has scarcely, as far as I know, introduced better plugin development
practices in the community.

# Implementation

The security manager and its supporting machinery have been totally
removed.