Removed MySQL connector
This is understandably a major change, but it is my understanding that
this was only added to help acclimate Bukkit developers to BungeeCord.
As LibertyCord targets a different audience, I feel that removing the
connector is justified. A lot of my experience includes dealing with
other DB systems (like PostgreSQL), so I feel strongly that I shouldn't
have to pay the cost to keep a driver I'll never use loaded in memory
all the time.