Began work on a new EventBus. This isn't done yet.
This resolves a major pain spot for me: handling events in an arbitrary
way. Currently you are limited to reflective method registration. This
means to dynamically generate a listener, you'll need to use reflection
to hook into the event bus itself (easily prone to breakage), or you
will need to dynamically generate classes at runtime (can be inefficient,
is prone to memory leaks, and is easy to get wrong).

This new event bus takes care of this issue. Its unit of granularity
is not the listener but the event executor. This makes it possible for
us to handle events any way we'd like to handle them.

The target use case I have in mind is integration with RxJava but I'm
sure someone will come up with a more creative use with time.